detalle BigDecimal

+++ proyecto_final/src/test/java/com/unimag/edu/proyecto_final/domine/repository/BaggageRepositoryTest.java
@@ -58,7 +58,7 @@ class BaggageRepositoryTest extends AbstractRepositoryIT {
                 .ticket(ticket)
                 .weight(12.3)
 
-                .fee(500.0)
+                .fee(BigDecimal.valueOf(500.0))
                 .tagCode("VGO001")
                 .build());
 
@@ -99,7 +99,7 @@ class BaggageRepositoryTest extends AbstractRepositoryIT {
         baggageRepository.save(Baggage.builder()
                 .ticket(ticket)
                 .weight(10.0)
-                .fee(0.0)
+                .fee(BigDecimal.valueOf(0.0))
                 .tagCode("BG002")
                 .build());
 
@@ -138,7 +138,7 @@ class BaggageRepositoryTest extends AbstractRepositoryIT {
         baggageRepository.save(Baggage.builder()
                 .ticket(ticket)
                 .weight(30.5)
-                .fee(1000.0)
+                .fee(BigDecimal.valueOf(1000.0))
                 .tagCode("BG003")
                 .build());
